requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
autopep8==1.5.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
# boto3==1.15.6
# botocore==1.18.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
dj-database-url==0.5.0
Django
django-ckeditor==6.2.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-froala-editor==4.0.15
django-heroku==0.3.1
django-jazzmin==2.4.8
django-js-asset==2.0.0
django-social-share==2.2.1
django-storages==1.10.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
flake8==4.0.1
gunicorn==20.0.4
h11==0.13.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
Jinja2==3.1.2
jmespath==0.10.0
Markdown==3.2.2
MarkupSafe==2.1.1
mccabe==0.6.1
numpy
openpyxl==3.0.10
outcome==1.1.0
pandas
Pillow==7.1.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2022.1
reportlab==3.5.49
requests==2.27.1
s3transfer==0.3.3
selenium==4.1.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.3.1
stripe==2.67.0
toml==0.10.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
# urllib3==1.26.9
webdriver-manager==3.5.4
webencodings==0.5.1
whitenoise==5.2.0
wsproto==1.1.0
xhtml2pdf==0.2.4
zipp==3.6.0
